Bugfixes for _mdr
Several bugs fixed:

1) Checking `len(training_features)` was looking at the number of
records, not features, which was not what was intended.

2) Added `.values` when passing data to the MDR transform method.
Otherwise it was passing a DataFrame, which is not what MDR accepts as
input.

3) Only allow 2-way MDR models, as the combinatorial space for 3-way
and beyond becomes too large too quickly. We will need to implement
some sort of built-in feature selection methods for the MDR function to
allow >2-way models.